changed bundling manager from cra to vite
