+++ client/src/App.css
@@ -1,0 +1,3 @@
+.App{
+  font-size: 1rem;
+}
