+++ client/src/StyleSheets/swiper-stars.css
